Fixed gtest flag in cmake Renamed entities folder to entity
