+++ CMakeLists.txt
@@ -29,6 +29,4 @@ if (BUILD_UNITTEST)
     message(STATUS "Unittest enabled")
     include_directories (external/gtest)
     add_subdirectory (external/gtest)
-    add_compile_options(-ftest-coverage -fprofile-arcs -fno-inline)
-    add_link_options(-fprofile-arcs -ftest-coverage -lgcov)
 endif()
