+++ domain/unittest/CMakeLists.txt
@@ -1,14 +1,12 @@
 project (domain_unittest)
 
+add_compile_options(-ftest-coverage -fprofile-arcs -fno-inline)
+set(CMAKE_EXE_LINKER_FLAGS "-fprofile-arcs -ftest-coverage -lgcov")
+
 add_executable(domain_unittest
     test_main.cpp
     test_auth.cpp
 )
-
-# add_library(gtest SHARED IMPORTED)
-# set_target_properties(gtest PROPERTIES
-#   IMPORTED_LOCATION "${CMAKE_SOURCE_DIR}/external/googletest/lib/libgtest.a"
-# )
 
 target_link_libraries (
     domain_unittest
